[workstation] 1. support 64 bit 2. use void * to store pointers 3. fix a bug where the size is stored as a char

+++ sym_backend/graph/Val.h
@@ -20,7 +20,7 @@ public:
     typedef unsigned short      BasicBlockIdType;
     typedef unsigned short      SymIDType;
     typedef unsigned int        ReadyType;
-    typedef unsigned int        IntValType;
+    typedef unsigned long int        IntValType;
     typedef unsigned char       ByteWidthType;
     typedef enum _ValType{
         ConstantIntValTy,
